Refactor 1.2: Split managers, models and serializers
Creating sub folders and moving each class/method to multiple files
