+++ app/repositories/managers/__init__.py
@@ -1,0 +1,6 @@
+from .base import BaseManager
+from .beverage import BeverageManager
+from .index import IndexManager
+from .ingredient import IngredientManager
+from .order import OrderManager
+from .size import SizeManager
