+++ app/repositories/managers/beverage.py
@@ -1,0 +1,16 @@
+from typing import Sequence
+
+from ..managers import BaseManager
+from ..models import Beverage
+from ..serializers import BeverageSerializer
+
+
+class BeverageManager(BaseManager):
+    model = Beverage
+    serializer = BeverageSerializer
+
+    @classmethod
+    def get_by_id_list(cls, ids: Sequence):
+        return cls.session.query(cls.model).filter(
+            cls.model._id.in_(set(ids))
+        ).all() or []
